Use NB-SpringBoot to access start.spring.io and generate a new project
Give the project a groupId and an artifactId as well as name and description. Furthermore, chose the first dependencies:

* web
* jpa
* devtools
* and h2 as a runtime dependency

By having h2 on the classpath, Spring Boot automatically configures an inmemory datasource for use during development and testing with JPA.

The devtools bring a reload server for web content as well as a modified class loader that supports changed class files. To enable a smooth debugging experience while using devetools, turn off "Apply code changes after save" for the Java debugger in NetBeans.