Add two more docker images
One is the database that will store our registration data in production. Wether to store that in a container is a good idea, I leave to the mind of the reader.

The other one is the container for our newly developed microservice.

Notice how those containers are linked together through the links element.

Also take note how PostgreSQL image is configured now: The global one is used for production and it gives default values to the username. Those replace the default postgres which we already used in the application-it.properties. The change takes also care of this.

The application container will start and immediatly shut down because the application fails to start. That is because we forgot to configure the database connection for Spring Boot.